backend/requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlmodel>=0.0.22
asyncpg>=0.30.0
greenlet>=3.1.0

# Authentication
pyjwt>=2.9.0

# Configuration
pydantic-settings>=2.6.0
python-dotenv>=1.0.0

# Development
httpx>=0.28.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
